Title: 컴퓨터와 함께하는 숫자 게임 2, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240821/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -4,7 +4,7 @@ public class Main {
 
     public static long M;
     public static int a;
-    public static int b;
+    public static long b;
 
     public static int minCount = 19;
     public static int maxCount = 0;
@@ -42,7 +42,7 @@ public class Main {
         M = sc.nextLong();
 
         a = sc.nextInt();
-        b = sc.nextInt();
+        b = sc.nextLong();
 
         for (int i = a; i <= b; i++) {
             find(i);
